backend/huggingface/requirements.txt
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
imutils==0.5.4
scikit-learn==1.3.2
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
